Fix server rendering error of battery icon
This was because I had misunderstood which files
are considered routes. Its not just 'special'
files like page and layout, its any file, so I
think the battery icon component was being
interpreted as a `/battery/[id]/battery-icon`
route.

Updated to put co-located files in folders
prefixed with an underscore as suggested in the
docs.