Use NextJS server actions
This should make it easier for candidates to refer
to the latest NextJS docs if they want to figure
out how it works as well as make the solution
'pure' rather than an unconventional mix of server
actions and React state.

Have also decided to reduce confusion by
having a single actions file. This is a common
convention in the docs and reduces needless
confusion from candidates who seem to mix up
which one of the two files they're looking at.